Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.8",
    "chromadb>=1.3.5",
    "fastapi>=0.124.0",
    "langchain>=1.1.3",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
